Consolidate filesystem-related functions.
