--- bork/file_finder.py
@@ -1,16 +1,0 @@
-from pathlib import Path
-
-
-def find_files(globs, service):
-    files = []
-
-    for glob in globs:
-        matches = map(str, Path().glob(glob))
-        files += list(matches)
-
-    print('Uploading to {}:'.format(service))
-    for filename in files:
-        print('- {}'.format(filename))
-    print('')
-
-    return files
