+++ bork/filesystem.py
@@ -1,0 +1,31 @@
+import configparser
+from pathlib import Path
+import shutil
+
+
+def load_setup_cfg():
+    setup_cfg = configparser.ConfigParser()
+    setup_cfg.read('setup.cfg')
+    return setup_cfg
+
+
+def find_files(globs, service):
+    files = []
+
+    for glob in globs:
+        matches = map(str, Path().glob(glob))
+        files += list(matches)
+
+    print('Uploading to {}:'.format(service))
+    for filename in files:
+        print('- {}'.format(filename))
+    print('')
+
+    return files
+
+
+def try_delete(path):
+    if Path(path).is_dir():
+        shutil.rmtree(path)
+    elif Path(path).exists():
+        raise Exception("{} is not a directory".format(path))
