+++ bork/github_uploader.py
@@ -1,8 +1,8 @@
-# from . import file_finder
+from .filesystem import find_files
 
 
 def upload(*globs, dry_run=False):
-    # files = file_finder.find_files(globs, 'GitHub')
+    # files = find_files(globs, 'GitHub')
 
     # if dry_run:
     #     print('NOTE: Skipping GitHub upload step since this is a dry run.')
