+++ bork/pypi_uploader.py
@@ -1,9 +1,9 @@
 from twine.cli import dispatch as twine_upload
-from . import file_finder
+from .filesystem import find_files
 
 
 def upload(*globs, dry_run=False):
-    files = file_finder.find_files(globs, 'PyPI')
+    files = find_files(globs, 'PyPI')
 
     if dry_run:
         print('NOTE: Skipping PyPI upload step since this is a dry run.')
